fix(agents-api): Move patching logic to litellm client
